refactor: Rename exceptions/ to exception/
