+++ src/winter/controllers/FrontController.java
@@ -18,13 +18,13 @@ import winter.data.Mapping;
 import winter.data.MappingMethod;
 import winter.data.ModelView;
 import winter.data.enums.RequestVerb;
-import winter.exceptions.AnnotationNotFoundException;
-import winter.exceptions.DuplicateMappingException;
-import winter.exceptions.InvalidPackageNameException;
-import winter.exceptions.InvalidRequestVerbException;
-import winter.exceptions.InvalidReturnTypeException;
-import winter.exceptions.MappingNotFoundException;
-import winter.exceptions.PackageProviderNotFoundException;
+import winter.exception.AnnotationNotFoundException;
+import winter.exception.DuplicateMappingException;
+import winter.exception.InvalidPackageNameException;
+import winter.exception.InvalidRequestVerbException;
+import winter.exception.InvalidReturnTypeException;
+import winter.exception.MappingNotFoundException;
+import winter.exception.PackageProviderNotFoundException;
 import winter.util.AnnotationScanner;
 import winter.util.ExceptionHandler;
 import winter.util.ReflectionUtil;
