+++ src/winter/exception/AnnotationNotFoundException.java
@@ -1,4 +1,4 @@
-package winter.exceptions;
+package winter.exception;
 
 public class AnnotationNotFoundException extends Exception {
     public AnnotationNotFoundException(String msg) {
