+++ src/winter/util/AnnotationScanner.java
@@ -11,9 +11,9 @@ import java.util.Map;
 import jakarta.servlet.ServletContext;
 import winter.data.Mapping;
 import winter.data.MappingMethod;
-import winter.exceptions.DuplicateMappingException;
-import winter.exceptions.InvalidPackageNameException;
-import winter.exceptions.PackageProviderNotFoundException;
+import winter.exception.DuplicateMappingException;
+import winter.exception.InvalidPackageNameException;
+import winter.exception.PackageProviderNotFoundException;
 import winter.utils.DirectoryScanner;
 import winter.utils.PackageNameValidator;
 import winter.utils.Utility;
